DevHub: Introduce a new template tag, `get_private_access_message()`, which generates a message for standalone functions or hooks marked with a private `@access` tag in their respective Docblocks.
If the Handbook plugin is available, more specifically if the `WPorg_Handbook_Callout_Boxes` class is available, the message markup is generated as an 'alert'-type callout box. Otherwise, fall-back markup and styles are applied to the message.

If an alternative function is referenced via an `@see` tag in the function or hook DocBlock, make a good-faith effort to try to link to the reference page for that element as an alternative within the access message text.

See #1498.