Support Forums: Cache lastpostmodified and sort by post id rather than post_modified_date.
The forums have enough traffic that running a sort by post_modified_gmt over all published posts is very expensive, and the cache is constantly invalidated by new topics and replies.

This uses an alternate query to find the date of the most recent topic or reply and uses that instead of a post_modified lookup. Instead of expiring the cached date when new topics or replies are added, the cached value is used for five minutes.