Handbook plugin: Highlights a menu link to the handbook home page when on any constituent handbook page.
Assuming the handbook page isn't already directly linked in the menu, preference is given to highlight a link to the front page of the current handbook. Barring the presence of such a link, it will check to see if there is a link to a 'handbook' or 'handbooks' page, which could be the case for multi-handbook sites.

Fixes #2283.