Official WordPress Events: Lossen loop condition to prevent infinite loop
The extra logging added in r5131 revealed that `$request_url` was `null`. It used to be an empty string, and still appears that way in the Meetup.com API console, but something must have changed. There were two changes to `json_decode()` in PHP 7.x which are related to empty strings and `null`, so that may have been the cause.