WordCamp: Load `wordcamporg` text domain based on user locale
Fixes an issue with WordCamp plugins' UIs still displaying in the site's
locale instead of the user's.

`get_user_locale()` was just introduced in WP 4.7, so it wasn't
available when this text domain strategy was implemented.

`get_user_locale()` falls back on `get_locale()` when necessary, so
it shouldn't cause a problem when a user doesn't have a specific locale
set.