Events: Replace object caching with stubs when running tests
Previously the object cache was enabled during tests, but it was bypassed by adding a cachebuster to the input. That was broken by r5552, because the cache keys are no longer built with the entire input. It wasn't immediately apparent, because the tests continued to pass as expected.

This commit changes that approach, so that the tests never load the object cache in the first place, and simply stub its functions instead.