WordCamp QBO: Bump class query limit to allow for growth.
The limit is just there as a sanity check, and `1000` is still low enough to serve that purpose. We'll probably have close to 150 WordCamps next year, so 300 might be too low in the next 5 years. We also already have close to 450 meetup groups, and might start creating classes for them in the foreseeable future.