WordCamp Coming Soon: Add Customizer field for custom intro message
This adds a new setting and control to the Coming Soon Page section
in the Customizer. Content added to the Custom Message textarea will
replace the default message shown on the Coming Soon page beneath the
header. This allows organizers a bit more flexibility in what they
communicate to site visitors before the site is officially launched.

Fixes #2901
Props Kau-Boy