WordCamp Post Type: Add note about number formatting
Sanitizing floats with variable comma and decimal point locations
(depending on the locale conventions of the user) turns out to be
quite a challenge. For today, we're going to take the easy way out
and just add a note to the input instructing to not use commas.

This also adds attributes to the number input field to make it more
compatible with a variable number of decimal points.