WP15: Use `Project-Id-Version` header to detect valid POMO content.
The value of the `Language` header is the shortest possible language code, which isn't readily available in the corresponding translation set, in order to compare against.

The `Project-Id-Version` header has a constant value, making it easier to compare against, and is still reliable enough for the purposes of smoke-testing the validity of the API response.

Props ocean90.
See https://wordpress.slack.com/archives/C02QB8GMM/p1521992879000021?thread_ts=1521943061.000060&cid=C02QB8GMM.