WordCamp tickets: Avoid permissions issue when generating email link
The notification emails for life-threatening allergies and special
accommodations are supposed to include a link to the attendee post, so the
organizer can find necessary details to follow up. However, using
`get_edit_post_link()` doesn't work when a non-site admin buys a ticket,
because the function checks if the current user can edit the post. This
generates the link manually instead.