WordCamp: Temporarily disable the privacy tools in 4.9.6
At this point the privacy tools in Core are only designed for single site
installations. Using them on a large multisite network like WordCamp.org could
have negative unintended consequences, especially since the user table is
shared across all of the sites. Once we have a system in place for handling
privacy requests on a network-wide basis, we can re-enable the tools.