WordCamp Speakers: Ensure REST API always returns avatar URLs
In the case that a speaker post did not have either a WordPress.org
username or a Gravatar Email for the speaker, the REST API response
would have empty fields for each of the avatar sizes instead of URLs
for the default avatar (e.g. the Mystery Person). This adds a check
on the contents of those fields and populates them with the default
URLs if necessary.