WordCamp Post Types: Include avatars in speaker endpoint for consistency
The Core setting controls whether or not avatars are displayed in comments,
which is not directly related to whether or not they should show up in the
API endpoint. There is other code like the Speakers block that relies on this
data always being available, and that would break if the setting were
disabled.

Props iandunn