WordCamp: Remove obsolete helper function
The `wcorg_register_meta_only_on_endpoint` function was created to handle
registering a post meta field for only a particular post type, before Core
had a way to do so with object subtypes. Now that Core supports this, we
have stopped using this function, and it can safely be removed.