Theme Directory: Start extracting headers from the `readme.txt` file in themes.
This is an MVP and does not use a proper readme parser as used in the Plugin Directory.
This is intended to collect the minimal header data we current have a need for.

In the future it's likely this will be expanded to allow more data from `readme.txt` to be utilised, but that's not the current priority.

See #3718