WC Docker: Prevent overwriting `rest_route` if json_route is null
JSON REST API plugin is depreciated and will not be there for new installations. This patch adds a conditional check so that in cases `json_route` is not present, core's API handlers will handle the requests.
Request with `json_route` will anyways fail when JSON REST API plugin is not installed, although this should be rare because it should be hardcoded only in some old posts.