WC Docker: Move WordPress files to `./mu` to match production.
In production environment, we have WordPress files in `public_html/mu` directory, so it makes sense to have them in same location in Docker environment as well.

This also required config change in `nginx` so that static files in `wp-content` are served properly given that root is changed.

This patch also has a minor fixup of http->https.