Official WordPress Events: Update source of Meetup_Client class
The WordCamp codebase has moved to GitHub, and the files in meta.svn are no
longer maintained (and will eventually be removed). There are some potential
pitfalls during deployment with using GitHub's svn bridge for svn externals,
so it has been decided to maintain a separate version of the Meetup_Client
instead. This also adds a helper method for loading required files and
instantiating the client class, since it needs to happen in multiple places
in this plugin.