Slack: Thread trac-bot ticket responses.
When a ticket is referenced within a thread, trac-bot will no longer send it to the global channel (often wtihout any context).

There's an option to reply in-thread, but also to broadcast to the wider channel, but thought that best to avoid for now.
Unfortunately the slack outgoing-webhook implementation doesn't announce whether the threaded message we're being triggered by was broadcast or not.

Fixes #2555.