Theme Directory: Don't use the global `$wp_query` object when querying themes.
This was causing cached results to be affected by other query params that `WP_Query` accepts through `$_GET`.

This will cause an increase in queries on same pages, but should be unnoticed thanks to the heavy caching we employ on these pages.

Debug left in to ignore caches with only 1 result, can be removed in a few hours after all caches expire.

Fixes #3215.