BuddyPress.org/bbPress.org: Add Survey support to bb-base header area.
This commit introduces a new template part (header-survey.php) and uses it across the 3 BuddyPress.org WordPress themes (Codex, Developer, and Base).

Because Surveys are done annually, it makes the most sense to leave this as a template part that can more easily be included, excluded, and modified in future years.

Props mercime.

Fixes #5208.