Plugin Directory: Search: Retry search queries when a API->ES timeout is encountered.
The volume of affected queries is low, but high enough that it's annoying on WordPress.org error logging.
This simply retries the query when the error rates are low (ie. ES is probably not unresponsive/overloaded, and the network isn't having issues) which means this shouldn't exacerbate any real errors.

Hopefully quietens the PHP Warnings related to Search timeouts, and fixes search results for the 1 in a million end-users who we didn't have a stale cache for.