Plugin Directory: Remove unused React code.
As part of the Plugin Directory refresh #1719, the UI was rewritten into a React app, but currently only the screenshot slider uses React. This removes all the React JS except for the screenshots, and flattens the Sass into less complex files. Additionally it updates the packages and JS build process to use @wordpress/scripts.

Fixes #5431.