Translate: Auto-correct spaced placeholders, common with machine translations.
This replaces `% 1 $ s` with `%1$s` in translations if it corrects the translation.

Machine translations aren't always wanted, but in many cases simply re-submitting the translation with the fixed placeholder is all that's done, this just reduces the amount of warnings generated, allowing translators to focus on the language content of the string.

See #5563, #5154.