Theme Directory: Fix theme uploads where new files were failing to be added/deleted when the theme added/deleted a file which included a space in the filename.
This caused some themes to fail to commit new files added when the theme contained a file named similar to `assets/js/custom.min copy.js`.
This would have also caused some theme uploads to fail to remove a file, if that file contained a space too.

Themes with new files which did not contain a space were unaffected.
Themes with files deleted which did not contain a space were unaffected.