Handling More Trademark Dramas
Clarifying the language that 'working around' the trademark catcher (i.e. using 'Wordpess' instead of 'WordPress' or 'Wh4tsApp' instead of 'WhatsApp') is not okay, AND that unless you're the trademark OWNER, it's probably not an error.

Adding in more trademarks per owner requests.

Adding in commonly used 'workarounds' because people are weird.

Fixes #5729