Plugin Directory: Remove "style" requirement when validating block plugins.
Previously, either `editorStyle` or `style` was required for block plugins, but this was not communicated to the plugin developer (only as a recommendation). This caused some block plugins to pass the validation in UI, but still fail validation behind the scenes. Omitting a style doesn't break anything on the core side, so it should be allowed.

Props oskosk for flagging an affected block.