Theme Directory: Add a set of checkboxes authors must agree to before uploads can be processed.
This ensures that authors are aware of the requirements prior to the upload being processed (and we can now assume they are aware of it).

The result of these checkboxes are not stored anywhere, but the upload cannot proceed unless they're checked, so all uploads after now will have agreed with them.

Fixes #1514.