Plugin Directory: Bypass the block plugin search when searching for core blocks.
If a core block (`core/*`) is disabled or missing from the editor, the missing block handler will try to look for it on WordPress.org, but no plugin should provide core blocks. This will short-circuit out of the block search and return zero results, to avoid issues if a plugin claims to provide the core block.

See #5910.