Add optional attributes to the wpcredits shortcode
Adds class and separator attributes to the shortcode callback function so that
the string of props names can better be styled to match the new design for the
WordPress.org News site.

The defaults for these attributes are set so that all existing uses of the
shortcode on the News site will be able to match the design without having to go
back and edit a lot of posts.