Trac: GitHub PRs: Use more specific matching for some PRs.
If the PR endpoint was called with a webhook payload which is for a repo that is not recognised by the endpoint, it would default to use core.trac, even if the ticket number was pulled from a non-core .trac.wordpress.org link.

This allows for the direct regex's to override the trac it ends up on when the link being matched is specific.

Also added a few other Repo's statically while I'm here.