Slack: Announce: Don' try to broadcast messages in private channels to the parent channel.
Some teams have private channels, which are in the format of `#team-privatechannel`, if there's any whitelisted users for the #team channel this would accidentally broadcast the private message publicly.
This has not been a problem in the past, as most private channels parent channels are also private (such as Security and Infrastructure related groups).

In an ideal world, we'd check to see if the channel status is the same (ie. if the parent channel was also private, we'd broadcast to that) but we don't have that information available in this handler.

See #4747.