Support Forums: Attempt to send notifications for new topics/replies when it's caught within the moderation or spam queue.
bbPress subscription notifications (and WordPress.org's) are hooked to the `bbp_new_topic` action, however they also verify that the thread is public before notifying.
Unfortunately, bbPress does not trigger the notifications upon the thread being unspammed/approved, and neither do the WordPress.org notifications.

This fixes this condition by triggering the "new topic created" hook after a topic/reply is unspammed/approved for the first time, causing the subscriptions codes to run as if it's a brand-new thread.

This should fix plugin subscription notifications not being sent to some plugin authors, and topic authors who are unaware of a reply because the responding author was caught in a moderation queue.

It's worth noting that replies still only trigger notifications (matchers, topic/forum subscriptions, @mention's, etc) if the topic is public AND the reply will be public.
If a reply is added to a pending topic, notifications for that reply will NOT be re-sent, unless the reply was also in a pending state unrelated to the topic and is then approved after the topic is.

See #3674.