WordPress.org Robots.txt: Correctly limit `/plugins/search/*` crawl limiting.
In [11722] I added blocking for the `plugins/search` endpoint, however, I did so by not including a trailing slash.
The result of this is that the rule has blocked access to plugins whose slugs begin with `search` (i.e. `/plugins/search-*/`).

By adding a trailing slash here, it correctly applies the Disallow rule to `/plugins/search/*` instead.

Props digamberpradhan for noticing/reporting the issue.
See #5323.
Fixes #6413.