Support Forums: Always require a reason for topic reports.
When reporting a topic, users have had the option to choose from a set of pre-determined reasons, or if none of the reasons were applicable, choosing the "other" alternative would give them a freeform text-field.

Using a pre-determined reason may leave reports short on context, where it could be down to the reporters interpretation why a reason was chosen, and would not be clear to those looking over reports.
This change addresses that by always requiring a user to not just choose a reason, but also provide context via the field asking "Why are you reporting this topic" when using the report tool in the sidebar.

Fixes #5706.