Plugin Directory: Admin: Link to any known previous emails related to the plugin.
This is powered through a similar way to the existing Helpscout plugins widget, instead of displaying the plugin related to the current email, it displays any emails that have ever been linked to a plugin.

This is not 100% accurate, as helpscout users will know, but it makes it a lot easier to find previous related emails.