Plugin Directory: Add a bin script to process email bounces in HelpScout.
This bin script can be run after bulk emails to automatically revoke commit or close plugins as required when the authors emails are bouncing.

Data is logged to allow plugin reviewers to see why an action was taken, and the Author Note (visible to plugin committers) explains the next steps if their plugin was closed.

See #6920.