Events: Only search Geonames for places with populations `> 500`
This ensures that searching for `Genova` will find the populated city with events, rather than the small village with no events.

It seems unlikely that they'll be a meetup in town with `< 500`, but we can revisit this if we run into that. The smallest place I saw with a meetup had a population of around `6,000`.